atm application

+++ src/Deposit.java
@@ -5,7 +5,7 @@ class Deposit extends Acc_balance{
         Scanner sc = new Scanner(System.in);
         System.out.print("How mush you deposit $.. ");
         float deposit_amount = sc.nextFloat();
-        float bal=getBalance();
+        double bal=getBalance();
         bal = deposit_amount + bal;
         System.out.println(bal);
         setBalance(bal);
